+++ src/dataset2D.h
@@ -34,12 +34,12 @@ public:
 	/// Constructor.
 	///
 	/// \param numberOfPoints Initial size of the buffers.
-	explicit Dataset2D(const unsigned int &numberOfPoints);
+	explicit Dataset2D(const size_t &numberOfPoints);
 
 	/// Resizes the internal buffers to hold the specified number of points.
 	///
 	/// \param numberOfPoints New size of internal buffers.
-	void Resize(const unsigned int &numberOfPoints);
+	void Resize(const size_t &numberOfPoints);
 
 	/// Reverses the order of the data stored in this object.
 	void Reverse();
@@ -54,7 +54,7 @@ public:
 
 	/// Gets the number of points stored in this object.
 	/// \returns The number of points stored in this object.
-	unsigned int GetNumberOfPoints() const { return mXData.size(); }
+	size_t GetNumberOfPoints() const { return mXData.size(); }
 
 	/// \name Private data accessors
 	/// @{
